Trb 44 without tests

+++ src/main/java/com/greenfox/tribes1/Building/Building.java
@@ -20,7 +20,7 @@ public abstract class Building {
   @GeneratedValue(strategy = GenerationType.IDENTITY)
   private Long id;
   private Long level;
-  private Long HP;
+  private Float HP;
   private Timestamp started_at;
   private Timestamp finished_at;
